v0/pyproject.toml
[project]
name = "advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.20",
    "google-adk==1.22.0",
    "google-cloud-aiplatform>=1.119.0",
    "google-cloud-firestore>=2.21.0",
    "pydantic>=2.11.9",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
]
